+++ common/src/main/java/io/novafoundation/nova/common/utils/stateMachine/list/states/EmptyProgressState.kt
@@ -1,5 +1,6 @@
 package io.novafoundation.nova.common.utils.stateMachine.list.states
 
+import io.novafoundation.nova.common.data.model.PageOffset
 import io.novafoundation.nova.common.data.model.hasNext
 import io.novafoundation.nova.common.utils.stateMachine.list.PaginatedListStateMachine
 import io.novafoundation.nova.common.utils.stateMachine.list.PaginatedListStateTransition
@@ -14,7 +15,7 @@ class EmptyProgressState<T>(query: String) : PaginatedListState<T>(query) {
                 event.newPage.isEmpty() -> emitState(EmptyState(query))
                 event.newPage.nextOffset.hasNext() -> emitState(DataState(query, event.newPage, event.newPage.nextOffset))
 
-                else -> fullDataLoaded(query, event.newPage)
+                else -> emitState(FullDataState(query, event.newPage))
             }
 
             is PaginatedListStateMachine.Event.PageError -> handlePageError(event, EmptyState(query))
@@ -23,6 +24,16 @@ class EmptyProgressState<T>(query: String) : PaginatedListState<T>(query) {
 
             is PaginatedListStateMachine.Event.QueryChanged -> handleQueryChanged(event)
         }
+    }
+
+    context(PaginatedListStateTransition<T>)
+    override suspend fun bootstrap() {
+        emitSideEffect(
+            PaginatedListStateMachine.SideEffect.LoadPage(
+                PageOffset.Loadable.FirstPage,
+                ""
+            )
+        )
     }
 
     override fun toString(): String {
