+++ feature-pay-impl/src/main/java/io/novafoundation/nova/feature_pay_impl/data/raise/auth/storage/RaiseAuthStorage.kt
@@ -6,7 +6,6 @@ import io.novafoundation.nova.common.data.secrets.v2.SecretStoreV2
 import io.novafoundation.nova.common.data.storage.encrypt.EncryptedPreferences
 import io.novafoundation.nova.common.utils.toStruct
 import io.novafoundation.nova.feature_account_api.data.secrets.keypair
-import io.novafoundation.nova.feature_account_api.domain.interfaces.AccountRepository
 import io.novafoundation.nova.feature_account_api.domain.model.MetaAccount
 import io.novasama.substrate_sdk_android.encrypt.keypair.substrate.Sr25519Keypair
 import io.novasama.substrate_sdk_android.scale.EncodableStruct
